chore: move token classes into chat templates
